Moving URL_QUERY from controller to core Moving file name queries.utils.js to query.utils.js and keeping in reports path.
