+++ webroot/js/common/core.app.js
@@ -125,7 +125,6 @@ function getCoreAppPaths(coreBaseDir, coreBuildDir, env) {
             'help-handler'                : coreWebDir + '/js/handlers/HelpHandler',
             'content-handler'             : coreWebDir + '/js/handlers/ContentHandler',
             'validation'                  : coreWebDir + '/assets/backbone/backbone-validation-amd',
-            'qe-utils'                    : coreWebDir + '/js/common/qe.utils',
             'mon-infra-node-list-model'   : coreWebDir + '/js/models/NodeListModel',
             'mon-infra-log-list-model'    : coreWebDir + '/js/models/LogListModel',
             'mon-infra-alert-list-view'   : coreWebDir + '/js/views/AlertListView',
@@ -445,7 +444,6 @@ var coreBundles = {
             'help-handler',
             'content-handler',
             'validation',
-            'qe-utils',
             'core-basedir/js/views/BarChartInfoView',
             'core-basedir/js/views/BreadcrumbDropdownView',
             'core-basedir/js/views/BreadcrumbTextView',
