+++ webroot/js/common/core.constants.js
@@ -405,6 +405,7 @@ define([
 
         this.BYTE_PREFIX = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
 
+        this.URL_QUERY = "/api/qe/query";
         this.URL_TABLES = "/api/qe/tables";
         this.URL_PREFIX_TABLE_SCHEMA = "/api/qe/table/schema/";
