+++ webroot/js/models/LogListModel.js
@@ -2,17 +2,20 @@
  * Copyright (c) 2015 Juniper Networks, Inc. All rights reserved.
  */
 
-define(['contrail-list-model'], function(ContrailListModel) {
+define([
+       'contrail-list-model',
+       'core-basedir/reports/qe/ui/js/common/qe.utils'
+], function(ContrailListModel, qeUtils) {
     var LogListModel = function() {
         var filter = "(Type = 1 AND Level <= 4) OR (Type = 10 AND " +
             "Level <= 4)";
         var qObj = {'table': 'MessageTable', 'level': 4, 'filter': filter,
             'limit': 10, 'minsSince': 10};
-        var postData = cowqu.formatQEUIQuery(qObj);
+        var postData = qeUtils.formatQEUIQuery(qObj);
         var listModelConfig = {
             remote : {
                 ajaxConfig : {
-                    url : '/api/qe/query',
+                    url : cowc.URL_QUERY,
                     type: 'POST',
                     data: JSON.stringify(postData)
                 },
